feat: restructure booking service
