--- src/services-booking-update/build.gradle.kts
@@ -1,35 +1,0 @@
-import java.text.SimpleDateFormat
-import java.util.Date
-
-plugins {
-    id("java-library")
-}
-
-description = "services-booking-update"
-group = "skladinya.services.booking-update"
-
-dependencies {
-    implementation(project(":domain"))
-
-    implementation("org.springframework:spring-context")
-    implementation("org.springframework:spring-web")
-}
-
-tasks.jar {
-    manifest {
-        var date = Date()
-        var javaVersion = System.getProperty("java.version")
-        var vmVendor = System.getProperty("java.vm.vendor")
-        var vmVersion = System.getProperty("java.vm.version")
-        attributes["Created-By"] = String.format("%s (%s %s)", javaVersion, vmVendor, vmVersion)
-        attributes["Gradle-Version"] = "Gradle $gradle.gradleVersion"
-        attributes["Build-Date"] = SimpleDateFormat("yyyy-MM-dd").format(date)
-        attributes["Build-Time"] = SimpleDateFormat("HH:mm:ss.SSSZ").format(date)
-        attributes["Build-By"] = "FRANK"
-        attributes["Bundle-Name"] = "SklaDinya Booking Update Service"
-        attributes["Bundle-SymbolicName"] = archiveBaseName
-    }
-    archiveBaseName.set("skladinya-services-booking-update")
-    archiveClassifier.set("")
-    archiveVersion.set("")
-}
