+++ src/settings.gradle.kts
@@ -9,7 +9,6 @@ include(":services-operator")
 include(":services-storage")
 include(":services-user")
 include(":services-booking")
-include(":services-booking-update")
 include(":services-payment")
 include(":services-price")
 include(":persistence-postgres")
